+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.6.0] - 2020-11-29
 ### Added
 - Added a new command 'remove' (placeholder)
+- Added a new command 'list' (placeholder)
 
 ## [0.5.1] - 2020-11-24
 ### Added